saveXml merged
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk@1375 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ wired/src/engine/Settings.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2004-2006 by Wired Team
+// Copyright (C) 2004-2007 by Wired Team
 // Under the GNU General Public License Version 2, June 1991
 
 #include "version.h"
@@ -92,11 +92,11 @@ void Settings::SetDefault()
   ConfDeprecated = false;
 }
 
-void Settings::ReadChannels(wxString Group, vector<long>& list)
+void Settings::ReadChannels(wxString Group, vector<int>& list)
 {
   wxString			s;
-  long				l;
-  long				val;
+  int				l;
+  int				val;
 
   conf->SetPath(Group);
   for (l = 0; ; l++)
@@ -108,11 +108,11 @@ void Settings::ReadChannels(wxString Group, vector<long>& list)
     }
 }
 
-void Settings::SaveChannels(wxString Group, vector<long>& list)
+void Settings::SaveChannels(wxString Group, vector<int>& list)
 {
-  vector<long>::iterator	i;
+  vector<int>::iterator	i;
   wxString			s;
-  long				l;
+  int				l;
 
   if (list.size() == 0)
     return;
